Swap to generic component
